Files actually removed

+++ src/main/java/com/mchg/tadremoelle/controllers/api/ImageController.java
@@ -44,7 +44,15 @@ public class ImageController {
 
     @DeleteMapping("/{id}")
     public void deleteImage(@PathVariable("id") Long id) {
-        imageService.deleteImage(id);
+        File f = new File(imageService.getImageById(id).getImageUrl());
+        if (f.delete()) {
+            System.out.println(f.getName() + " deleted");
+            imageService.deleteImage(id);
+        }
+        else {
+            System.out.println("File deletion failed");
+        }
+
     }
 
     @RequestMapping(value = "/uploadFile", method = RequestMethod.POST)
